requirements.txt
Flask==2.2.5
python-dotenv==0.21.1
requests==2.32.3
